feat: onboard provisions the flows schema (#246)
In provisioned mode (manage-schema=false) the collector only validates the
schema, while onboard's ensureShared opened with GRANT/ALTER on <db>.flows —
both requiring the table to exist. Nobody created it, so a fresh provisioned
deployment could not be onboarded (chicken-and-egg).

Extract the flows table + samples view DDL into a neutral, database-qualified
org.riptide.schema.FlowsSchema shared by the collector's manage path and by
onboard, so the two definitions cannot drift. Prepend CREATE DATABASE +
CREATE TABLE to ensureShared, before the first GRANT/ALTER that needs the
table. The collector switches its unqualified DDL to the shared qualified form
(behavior-preserving on its pinned client).

onboard deliberately does not create the samples view: in provisioned mode
flow_reader is granted SELECT on flows and system.* only, so a samples view
would be inert. Exposing it to readers is a separate, spike-verified follow-up
(row policy propagates through the parameterized view under the default INVOKER
security; DEFINER would leak).

TenantOnboardingIT now onboards into an empty database, proving the table
bootstrap end to end against real ClickHouse.
