+++ src/main/java/org/riptide/provisioning/ProvisioningDdl.java
@@ -5,6 +5,8 @@
 
 package org.riptide.provisioning;
 
+import org.riptide.schema.FlowsSchema;
+
 import java.util.List;
 
 /**
@@ -12,9 +14,9 @@ import java.util.List;
  * whole recipe is one auditable place and lifts cleanly into a future {@code riptide-admin} module.
  *
  * <p>The model puts every identical-per-tenant part into one-time objects ({@link #ensureShared})
- * — the {@code flow_writer}/{@code flow_reader} roles carry the grants and the reader hardening,
- * and a single quota is keyed by user — so {@link #onboardTenant} reduces to the scoped users, two
- * role grants, and one literal row policy. All statements are idempotent
+ * — the {@code flows} schema, the {@code flow_writer}/{@code flow_reader} roles carrying the grants
+ * and the reader hardening, and a single quota keyed by user — so {@link #onboardTenant} reduces to
+ * the scoped users, two role grants, and one literal row policy. All statements are idempotent
  * ({@code IF NOT EXISTS} / {@code ALTER ROLE SETTINGS}), verified on ClickHouse 25.3.
  *
  * <p>Tenant/org names are validated ({@link TenantSpec}) to a safe charset; generated identifiers
@@ -25,10 +27,18 @@ public final class ProvisioningDdl {
     private ProvisioningDdl() {
     }
 
-    /** One-time shared objects: the two roles, the reader hardening, the CHECK barrier, the quota. */
+    /**
+     * One-time shared objects: the database and {@code flows} table (so onboarding a fresh
+     * provisioned deployment is self-sufficient — the collector only validates in
+     * {@code manage-schema=false} mode), then the two roles, the reader hardening, the CHECK
+     * barrier, and the quota. The schema DDL comes first because the {@code GRANT INSERT} and
+     * {@code ALTER TABLE … ADD CONSTRAINT} below require the table to exist.
+     */
     public static List<String> ensureShared(final String database, final long quotaBytes) {
         final String flows = ident(database) + ".flows";
         return List.of(
+                FlowsSchema.createDatabase(database),
+                FlowsSchema.createFlowsTable(database),
                 "CREATE ROLE IF NOT EXISTS flow_writer",
                 "GRANT INSERT ON " + flows + " TO flow_writer",
                 "CREATE ROLE IF NOT EXISTS flow_reader",
