+++ src/test/java/org/riptide/schema/FlowsSchemaTest.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2026 Riptide Labs, <https://github.com/Riptide-Labs>
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+package org.riptide.schema;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * The flow schema DDL is the single source shared by the collector's manage path and {@code
+ * onboard}, so the load-bearing properties are that it is database-qualified (works on an unpinned
+ * client), idempotent, and that the {@code samples} view still references the same qualified table.
+ */
+class FlowsSchemaTest {
+
+    @Test
+    void createDatabaseIsIdempotentAndQuoted() {
+        assertThat(FlowsSchema.createDatabase("riptide"))
+                .isEqualTo("CREATE DATABASE IF NOT EXISTS `riptide`");
+    }
+
+    @Test
+    void createFlowsTableIsQualifiedAndIdempotent() {
+        final String ddl = FlowsSchema.createFlowsTable("riptide");
+        assertThat(ddl.strip()).startsWith("CREATE TABLE IF NOT EXISTS `riptide`.flows (");
+        // A representative column and the engine/partitioning survive the extraction unchanged.
+        assertThat(ddl)
+                .contains("tenant String,")
+                .contains("clockCorrection Nullable(Int64)")
+                .contains("ENGINE = MergeTree()")
+                .contains("PARTITION BY toYYYYMMDD(timestamp)");
+    }
+
+    @Test
+    void createSamplesViewQualifiesBothViewAndSourceTable() {
+        final String ddl = FlowsSchema.createSamplesView("riptide");
+        assertThat(ddl.strip()).startsWith("CREATE OR REPLACE VIEW `riptide`.samples AS");
+        assertThat(ddl).contains("FROM `riptide`.flows AS flow");
+        // The view parameter is a literal placeholder bound at SELECT time, not at CREATE time.
+        assertThat(ddl).contains("{ival:Int64}");
+    }
+
+    @Test
+    void qualifiesToTheGivenDatabase() {
+        assertThat(FlowsSchema.createFlowsTable("acme_prod"))
+                .contains("CREATE TABLE IF NOT EXISTS `acme_prod`.flows (");
+        assertThat(FlowsSchema.createSamplesView("acme_prod"))
+                .contains("VIEW `acme_prod`.samples AS")
+                .contains("FROM `acme_prod`.flows AS flow");
+    }
+}
